Bug Fixed
This bug fixes the issue when an expression is deleted but the error plot shows other selected plot if it exists else gives could not find expression statement. Earlier if any selected expression was deleted it just gave that  'could not find expression with id 0', now it only gives when all expressions are deleted.